Hey,

I was doing some build-time profiling on a project and tried `cargo llvm-lines` which helps visualizing the amount of code generated by the monomorphization of generic functions. Doing so, I saw that some of the biggest contributors were functions from `tonic`.

For some of them it is possible de reduce the amount of monomophization quite easily.

This is the result of my iteration on the biggest contributor which was refactorable. It should help for smaller binaries and faster build time.

Output of `cargo llvm-lines` on my project, before:
```
Lines           Copies        Function name 
 -----           ------        ------------- 
30754 (1.2%)     17 (0.0%)  <tonic::codec::decode::Streaming<T> as futures_core::stream::Stream>::poll_next
13977 (0.5%)     17 (0.0%)  tonic::codec::decode::Streaming<T>::decode_chunk
 7158 (0.3%)     14 (0.0%)  tonic::codec::decode::Streaming<T>::trailers::{{closure}}
  742 (0.0%)     42 (0.1%)  tonic::codec::decode::Streaming<T>::trailers::{{closure}}::{{closure}}
  210 (0.0%)     14 (0.0%)  tonic::codec::decode::Streaming<T>::trailers
```

And after:
```
Lines           Copies        Function name 
 -----           ------        ------------- 
6339 (0.3%)     17 (0.0%)  <tonic::codec::decode::Streaming<T> as futures_core::stream::Stream>::poll_next
5730 (0.2%)     14 (0.0%)  tonic::codec::decode::Streaming<T>::trailers::{{closure}}
4952 (0.2%)     17 (0.0%)  tonic::codec::decode::Streaming<T>::decode_chunk
 221 (0.0%)      1 (0.0%)  tonic::codec::decode::StreamingInner::trailers::{{closure}}
 210 (0.0%)     14 (0.0%)  tonic::codec::decode::Streaming<T>::trailers
  53 (0.0%)      3 (0.0%)  tonic::codec::decode::StreamingInner::trailers::{{closure}}::{{closure}}
```

This gives an about 35_336 llvm-lines reduction