@LucioFranco 2021-12-08
This seems really cool! Do you have any compile time benchmarks by chance? Curious what the real impact is.