@Guiguiprim 2021-12-09
This awkward moment when the benchmark decided not to go your way.

With the attached minimal project:

Before this commit 
```bash
> cargo b --release
> stat target/release/minimal-tonic-monomorphization -c "%s"
7_463_512

> hyperfine --prepare "rm -rf target/release/build/minimal-tonic-monomorphization-*" "cargo b --release"
Benchmark #1: cargo b --release
  Time (mean ± σ):     12.133 s ±  0.185 s    [User: 48.109 s, System: 0.762 s]
  Range (min … max):   11.855 s … 12.439 s    10 runs
```

After this commit
```bash
> cargo b --release
> stat target/release/minimal-tonic-monomorphization -c "%s"
7_451_720

> hyperfine --prepare "rm -rf target/release/build/minimal-tonic-monomorphization-*" "cargo b --release"
Benchmark #1: cargo b --release
  Time (mean ± σ):     12.670 s ±  0.178 s    [User: 48.070 s, System: 0.744 s]
  Range (min … max):   12.530 s … 13.127 s    10 runs
```

The binary did decreased by 12ko, but the build time appears a bit longer. Which feels weird.

[minimal-tonic-monomorphization.tar.gz](https://github.com/hyperium/tonic/files/7685731/minimal-tonic-monomorphization.tar.gz)